+++ src/onegov.ballot/onegov/ballot/upgrade.py
@@ -3,6 +3,7 @@ upgraded on the server. See :class:`onegov.core.upgrade.upgrade_task`.
 
 """
 from onegov.core.upgrade import upgrade_task
+from sqlalchemy import Column, Text
 
 
 @upgrade_task('Rename yays to yeas', always_run=True)
@@ -13,3 +14,8 @@ def rename_yays_to_yeas(context):
     else:
         context.operations.alter_column(
             'ballot_results', 'yays', new_column_name='yeas')
+
+
+@upgrade_task('Add shortcode column')
+def add_shortcode_column(context):
+    context.operations.add_column('votes', Column('shortcode', Text()))